Change dimensions of zp to prevent segfaults. Various changes to silence warnings (not properly tested).

+++ analysis/phiexamine.f
@@ -283,7 +283,7 @@
          do k=1,nc
             cl(k)=(first+k*delta)
          enddo
-         nc=abs(2.*phimax)/abs(delta)
+         nc=int(abs(2.*phimax)/abs(delta))
 !         write(*,*)' Phimax', phimax,' fac10',fac10
 !         write(*,*)' Contours',nc,(cl(k),k=1,nc)
          iconsw=1+64
